LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/mutation-observer-expected.txt


PASS [Connected move] MutationObserver removal + insertion is tracked by moveBefore()
PASS [Disconnected move] MutationObserver removal + insertion is tracked by moveBefore()
